Fix integration tests for previous release.
